+++ provider/provider.go
@@ -17,6 +17,8 @@ import (
 )
 
 const esKeyName = "elasticsearch"
+const kbKeyName = "kibana"
+const fleetKeyName = "fleet"
 
 func init() {
 	// Set descriptions to support markdown syntax, this will be used in document generation
@@ -27,9 +29,9 @@ func init() {
 func New(version string) *schema.Provider {
 	p := &schema.Provider{
 		Schema: map[string]*schema.Schema{
-			esKeyName: providerSchema.GetEsConnectionSchema(esKeyName, true),
-			"kibana":  providerSchema.GetKibanaConnectionSchema(),
-			"fleet":   providerSchema.GetFleetConnectionSchema(),
+			esKeyName:    providerSchema.GetEsConnectionSchema(esKeyName, true),
+			kbKeyName:    providerSchema.GetKibanaConnectionSchema(),
+			fleetKeyName: providerSchema.GetFleetConnectionSchema(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"elasticstack_elasticsearch_ingest_processor_append":            ingest.DataSourceProcessorAppend(),
@@ -113,7 +115,7 @@ func New(version string) *schema.Provider {
 		},
 	}
 
-	p.ConfigureContextFunc = clients.NewApiClientFunc(version)
+	p.ConfigureContextFunc = clients.NewApiClientFuncFromSDK(version)
 
 	return p
 }